Use virtual only if length greater than
